Added check for legal entity
Added a check to see if the account legal entity has already been added
to the account, if it has then no action is taken. Also catch a
DbUpdateException for if the constraint has been caught on a duplicate
entry being added. Not testable due to the type of exception it is
catching as the base exception